feat: improve Handlers
