+++ packages/core/src/events/index.tsx
@@ -1,5 +1,6 @@
 export { useEventHandler } from './EventContext';
 export * from './CoreEventHandlers';
 export * from './DefaultEventHandlers';
+export * from './Handlers';
 export { Events } from './Events';
 export * from './defineEventListener';
